completed back-end and go to develop front-end by nuxtjs

+++ src/tasks/tasks.service.ts
@@ -1,5 +1,6 @@
 import { BadRequestException, Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
+import { User } from 'src/auth/user.entity';
 import { CreateTaskDto } from './dto/create-task.dto';
 import { GetTaskFilterDto } from './dto/get-task-filter.dto';
 import { TaskStatus } from './task-status.enum';
@@ -10,8 +11,8 @@ import { TaskRepository } from './task.repository';
 export class TasksService {
     constructor(@InjectRepository(TaskRepository) private taskRepository: TaskRepository) {}
 
-    async getTaskById(id: number) : Promise<Task> {
-        const found = await this.taskRepository.findOne(id);
+    async getTaskById(id: number, user: User) : Promise<Task> {
+        const found = await this.taskRepository.findOne({where: {id, userId: user.id}});
         
         if(!found) {
             throw new NotFoundException(`Task with ID ${id} not found`);
@@ -20,33 +21,20 @@ export class TasksService {
         return found;
     }
 
-    async createTask(createTaskDto: CreateTaskDto) : Promise<Task> {
-        const {title, description} = createTaskDto;
-
-        const task = new Task();
-        task.title = title;
-        task.description = description;
-        task.status = TaskStatus.OPEN;  
-
-        let taskCreated = await task.save();
-
-        if(!taskCreated) {
-            throw new BadRequestException('Can not insert task to database');
-        }
-
-        return task
+    async createTask(createTaskDto: CreateTaskDto, user: User) : Promise<Task> {
+        return this.taskRepository.createTask(createTaskDto, user);
     }
 
-    async deleteTaskById(id: number) : Promise<void> {
-        let execute = await this.taskRepository.delete(id);
+    async deleteTaskById(id: number, user: User) : Promise<void> {
+        let execute = await this.taskRepository.delete({id, userId: user.id});
 
         if(execute.affected === 0) {
             throw new NotFoundException(`Task with ID ${id} not found`);
         }
     }
     
-    async updateStatusTask(id: number, status: TaskStatus) : Promise<Task> {
-        let task = await this.getTaskById(id);
+    async updateStatusTask(id: number, status: TaskStatus, user: User) : Promise<Task> {
+        let task = await this.getTaskById(id, user);
 
         task.status = status;
         task.save();
@@ -54,24 +42,7 @@ export class TasksService {
         return task;
     }
 
-    async getTasks(filterDto: GetTaskFilterDto) : Promise<Task[]> {
-        return await this.taskRepository.getTasks(filterDto);
+    async getTasks(filterDto: GetTaskFilterDto, user : User) : Promise<Task[]> {
+        return await this.taskRepository.getTasks(filterDto, user);
     }
-    // async getTaskWithFilters(filterDto: GetTaskFilterDto) : Promise<Task[]> {
-    //     const {status, search} = filterDto;
-
-    //     let tasks = await this.getAllTasks();
-
-    //     if(status) {
-    //         tasks = tasks.filter(el => el.status === status);
-    //     }
-
-    //     if(search) {
-    //         tasks = tasks.filter(el => {
-    //            return  el.title.includes(search) || el.description.includes(search)
-    //         });
-    //     }
-
-    //     return tasks;
-    // }
 }
